// petforest-fe-cms/src/utils/enums/manager/manager.enum.ts
export enum MANAGER_USER_CONTENT {
  USER_LIST = 'USER_LIST',
  RESIGNED_LIST = 'RESIGNED_LIST',
}

export enum MANAGER_BRANCH_CONTENT {
  BRANCH_LIST = 'BRANCH_LIST',
  DELETE_LIST = 'DELETE_LIST',
}

export enum USER_BRANCH_TYPE {
  MAIN = 'MAIN',
  NORMAL = 'NORMAL',
}

export enum USER_PERMISSION {
  MODIFY = 'MODIFY',
  VIEW = 'VIEW',
  NOT_ALLOW = 'NOT_ALLOW',
}

export enum USER_RECEIVE_RESERVATION_PERMISSION {
  NOT_ALLOW = 'NOT_ALLOW',
  MODIFY = 'MODIFY',
}

export enum USER_POSITION {
  FUNERAL_MANAGER = 'FUNERAL_MANAGER',
}

export enum RESERVATION_TIME_SHIFT {
  AM = 'AM',
  PM = 'PM',
}

export enum RESERVATION_TIME_STATUS {
  OPENING = 'OPENING',
  CLOSED = 'CLOSED',
  COMPLETED = 'COMPLETED',
}